+++ src/Controller/Admin/OrderMain.php
@@ -13,7 +13,7 @@ use OxidEsales\Eshop\Core\Exception\StandardException;
 use OxidEsales\Eshop\Core\Registry;
 use OxidSolutionCatalysts\Unzer\Model\Payment;
 use OxidSolutionCatalysts\Unzer\Service\Payment as PaymentService;
-use OxidSolutionCatalysts\Unzer\Model\Order;
+use OxidEsales\Eshop\Application\Model\Order;
 use OxidSolutionCatalysts\Unzer\Traits\ServiceContainer;
 use UnzerSDK\Exceptions\UnzerApiException;
 
@@ -44,10 +44,6 @@ class OrderMain extends OrderMain_parent
         }
     }
 
-    /**
-     * @param Order $oOrder
-     * @return void
-     */
     public function sendShipmentNotification(Order $oOrder): void
     {
         $paymentService = $this->getServiceFromContainer(PaymentService::class);