+++ src/Controller/DispatcherController.php
@@ -14,6 +14,7 @@ use OxidEsales\Eshop\Application\Controller\FrontendController;
 use OxidEsales\Eshop\Application\Model\Order;
 use OxidEsales\Eshop\Core\Registry;
 use OxidSolutionCatalysts\Unzer\Model\TmpOrder;
+use OxidSolutionCatalysts\Unzer\Service\FlexibleSerializer;
 use OxidSolutionCatalysts\Unzer\Service\Transaction;
 use OxidSolutionCatalysts\Unzer\Service\Translator;
 use OxidSolutionCatalysts\Unzer\Service\UnzerSDKLoader;
@@ -35,12 +36,16 @@ class DispatcherController extends FrontendController
     public function __construct()
     {
         parent::__construct();
-        $this->transaction = $this->getServiceFromContainer(Transaction::class);
-        $this->unzerWebhooks = $this->getServiceFromContainer(UnzerWebhooks::class);
-        $this->unzerSDKLoader = $this->getServiceFromContainer(UnzerSDKLoader::class);
-        $this->translator = $this->getServiceFromContainer(Translator::class);
+        $this->transaction = $this->getTransactionService();
+        $this->unzerWebhooks = $this->getUnzerWebhooks();
+        $this->unzerSDKLoader = $this->getUnzerSdkLoader();
+        $this->translator = $this->getTranslator();
     }
 
+    /**
+     * @throws \UnzerSDK\Exceptions\UnzerApiException
+     * @throws \OxidSolutionCatalysts\Unzer\Exception\UnzerException
+     */
     public function updatePaymentTransStatus(): void
     {
         $jsonRequest = file_get_contents('php://input');
@@ -78,7 +83,7 @@ class DispatcherController extends FrontendController
         }
 
         $unzer = $this->unzerSDKLoader->getUnzerSDKbyKey($unzerKey);
-        $resource = $unzer->fetchResourceFromEvent((string)$jsonRequest);
+        $resource = $unzer->fetchResourceFromEvent($jsonRequest);
         $paymentId = $resource->getId();
 
         if ($paymentId) {
@@ -103,12 +108,12 @@ class DispatcherController extends FrontendController
         $order = oxNew(Order::class);
         $data = $this->transaction->getTransactionDataByPaymentId($paymentId);
 
-        if (!is_array($data) || !isset($data[0]['OXORDERID'])) {
+        if (!is_array($data) || !isset($data[0]['oxorderid'])) {
             return "Invalid Order Data";
         }
 
         $unzerPayment = $unzer->fetchPayment($paymentId);
-        if ($order->load($data[0]['OXORDERID'])) {
+        if ($order->load($data[0]['oxorderid'])) {
             return $this->updateOrder($order, $unzerPayment, $paymentId);
         }
 
@@ -172,6 +177,79 @@ class DispatcherController extends FrontendController
         return $this->translator->translate('oscunzer_TRANSACTION_NOTHINGTODO') . $paymentId;
     }
 
+    /**
+     * @throws \Exception
+     */
+    public function finalizeTmpOrder(
+        Payment $unzerPayment,
+        TmpOrder $tmpOrder,
+        array $tmpData,
+        bool $bError
+    ): string {
+        $return = $this->returnError();
+        if ($tmpOrder->load($tmpData['oxid'])) {
+            $flexibleSerializer = $this->getFlexibleSerializer();
+
+            /** @var Order $oOrder */
+            $oOrder = $flexibleSerializer->safeUnserialize(
+                $tmpData['tmporder'],
+                [
+                    Order::class
+                ]
+            );
+
+            if ($oOrder instanceof \stdClass || $oOrder === false) {
+                $oOrder = $flexibleSerializer->restoreOrderFromStrClass($tmpData['tmporder']);
+            }
+
+            if (!is_null($oOrder) && method_exists($oOrder, 'finalizeTmpOrder')) {
+                /** @var \OxidSolutionCatalysts\Unzer\Model\Order $oOrder */
+                $oOrder->finalizeTmpOrder($unzerPayment, $bError);
+                $tmpOrder->assign(['status' => 'FINISHED']);
+                $tmpOrder->save();
+
+                $return = $this->returnSuccess();
+            }
+        }
+
+        return $return;
+    }
+
+    protected function getFlexibleSerializer(): FlexibleSerializer
+    {
+        return $this->getServiceFromContainer(FlexibleSerializer::class);
+    }
+
+    protected function getUnzerSdkLoader(): UnzerSDKLoader
+    {
+        return $this->getServiceFromContainer(UnzerSDKLoader::class);
+    }
+
+    protected function getTranslator(): Translator
+    {
+        return $this->getServiceFromContainer(Translator::class);
+    }
+
+    protected function getUnzerWebhooks(): UnzerWebhooks
+    {
+        return $this->getServiceFromContainer(UnzerWebhooks::class);
+    }
+
+    protected function getTransactionService(): Transaction
+    {
+        return $this->getServiceFromContainer(Transaction::class);
+    }
+
+    protected function returnError(): string
+    {
+        return $this->translator->translate('oscunzer_ERROR_HANDLE_TMP_ORDER');
+    }
+
+    protected function returnSuccess(): string
+    {
+        return $this->translator->translate('oscunzer_SUCCESS_HANDLE_TMP_ORDER');
+    }
+
     private function handleTmpOrder(Payment $unzerPayment): string
     {
         $tmpOrder = oxNew(TmpOrder::class);
@@ -179,8 +257,8 @@ class DispatcherController extends FrontendController
         $tmpData = $tmpOrder->getTmpOrderByUnzerId($orderId);
 
         if (
-            isset($tmpData['OXID']) &&
-            $tmpOrder->load($tmpData['OXID']) &&
+            isset($tmpData['oxid']) &&
+            $tmpOrder->load($tmpData['oxid']) &&
             $this->hasExceededTimeLimit($tmpOrder)
         ) {
             $bError = !(
@@ -190,27 +268,6 @@ class DispatcherController extends FrontendController
             );
 
             return $this->finalizeTmpOrder($unzerPayment, $tmpOrder, $tmpData, $bError);
-        }
-
-        return $this->translator->translate('oscunzer_ERROR_HANDLE_TMP_ORDER');
-    }
-
-    private function finalizeTmpOrder(
-        Payment $unzerPayment,
-        TmpOrder $tmpOrder,
-        array $tmpData,
-        bool $bError
-    ): string {
-        if ($tmpOrder->load($tmpData['OXID'])) {
-            $aOrderData = unserialize(base64_decode($tmpData['TMPORDER']), ['allowed_classes' => [Order::class]]);
-            $oOrder = is_array($aOrderData) && isset($aOrderData['order']) ? $aOrderData['order'] : null;
-
-            if ($oOrder) {
-                $oOrder->finalizeTmpOrder($unzerPayment, $bError);
-                $tmpOrder->assign(['status' => 'FINISHED']);
-                $tmpOrder->save();
-                return $this->translator->translate('oscunzer_SUCCESS_HANDLE_TMP_ORDER');
-            }
         }
 
         return $this->translator->translate('oscunzer_ERROR_HANDLE_TMP_ORDER');