+++ tests/PhpStan/phpstan-bootstrap.php
@@ -53,11 +53,6 @@ class_alias(
 
 class_alias(
     \OxidEsales\Eshop\Application\Model\Order::class,
-    \OxidEsales\Eshop\Application\Model\Order_parent::class,
-);
-
-class_alias(
-    \OxidEsales\Eshop\Application\Model\Order::class,
     \OxidSolutionCatalysts\Unzer\Model\Order_parent::class,
 );